suite correction

+++ correction-todo-list/src/main/java/com/example/correctiontodolist/controller/TodoUrgentController.java
@@ -1,6 +1,7 @@
 package com.example.correctiontodolist.controller;
 
 import com.example.correctiontodolist.service.UrgentTodoService;
+import com.example.correctiontodolist.service.UserTodoService;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -21,27 +22,44 @@ public class TodoUrgentController {
     
     @Autowired
     private UrgentTodoService _urgentTodoService;
+
+    @Autowired
+    private UserTodoService _userTodoService;
     
     @GetMapping("add/{id}")
     @ResponseBody
     public void addToUrgent(@PathVariable("id") int id) throws IOException {
-        if(_urgentTodoService.addToUrgent(id)) {
-            _response.sendRedirect("/todos-html/detail/"+id);
+        if(_userTodoService.isLogged()) {
+            if(_urgentTodoService.addToUrgent(id)) {
+                _response.sendRedirect("/todos-html/detail/"+id);
+            }
+            //Redirection vers une page d'erreur si la todo ne peut pas être ajouter
         }
-        //Redirection vers une page d'erreur si la todo ne peut pas être ajouter
+        else {
+            _response.sendRedirect("/user/login");
+        }
     }
 
     @GetMapping("remove/{id}")
     @ResponseBody
     public void removeToUrgent(@PathVariable("id") int id) throws IOException {
-        if(_urgentTodoService.removeFromUrgent(id)) {
-            _response.sendRedirect("/todos-html/detail/"+id);
+        if(_userTodoService.isLogged()) {
+            if(_urgentTodoService.removeFromUrgent(id)) {
+                _response.sendRedirect("/todos-html/detail/"+id);
+            }
+            //Redirection vers une page d'erreur si la todo ne peut pas être ajouter
         }
-        //Redirection vers une page d'erreur si la todo ne peut pas être ajouter
+        else {
+            _response.sendRedirect("/user/login");
+        }
     }
 
     @GetMapping("")
-    public ModelAndView displayUrgent() {
+    public ModelAndView displayUrgent() throws IOException {
+        if(!_userTodoService.isLogged()) {
+            _response.sendRedirect("/user/login");
+
+        }
         ModelAndView vm = new ModelAndView("todos");
         vm.addObject("todos", _urgentTodoService.getUrgentsTodo());
         return vm;
